Ddd a new leaf node type with different magic to support larger values
for block storage and most other purposes, a 64kiB limit on value sizes
usually works fine, but when using in "kv" (key-value) mode, it it
sometimes beneficial to support larger values. To remain compatible, we
look at the magic cookie in simpletree meta-nodes and potentially decide
to use larger leaf node entries with more bits for value etc. This is
is only enabled when running through "kv", and the code remains
backwards-compatible with existing trees.