support flush and updating of .swap file on disk without unmount/shutdown.
This should make it more usable for day-to-day storage uses, and not
just for format-mount-copy-unmount style scenarios that are great for VM
and container images, but less so for cases where you happen to care
about the data being generated during prolonged use, without wanting
to unmount to, e.g., make a backup.

Write-back caching, which randomly reorders writes by location instead
of temporally, is still in place, so clean unmounts can still lead to
inconsistent file system state in some cases. The simplest fix for this
will be to delay committing data structure root updates until everything
in cache was been inserted into the dubtree structure. Only thing
missing to make this work is support for delayed deletion of unrefenced
chunks pending a data structure root commit.