trigger snapshot creation with SIGUSR1
sending this signal to the main process' pid will cause hardlinked
backup snapshot of the top-level swapdata directory. This snapshot can
be used as a separate virtual disk with its own UUID, or as a backup
that can be copied to remote storage. IOs are held back while the
snapshot is in progress.

XXX suspend io